backend/新建文件夹/requirements.txt
Django>=2.1.9
django-cors-headers==2.5.2
djangorestframework==3.9.2
djangorestframework-simplejwt==4.1.0
djangorestframework-stubs==0.4.2
django-rest-swagger==2.2.0
py
pbr==5.1.3
PyJWT==1.7.1
pytz==2018.9
six==1.12.0
stevedore==1.30.1
virtualenv==16.4.3
virtualenv-clone==0.5.1
virtualenvwrapper==4.8.4
dwebsocket==0.5.12
celery==4.4.2
django-celery==3.3.1
redis==2.10.6
django-redis==4.10.0


